place the function responsible to get the bank name in the parent component, fix some texts for inquiry and also include bank and and remove an extra item, fix some tests
